Función mutación mejorada e incorporada a los métodos de evolución

+++ Especimen.hpp
@@ -32,7 +32,7 @@ public:
     
     friend void cruce2Puntos(Especimen &padreA, Especimen &padreB,int minimo = 1,int maximo = 0);
     friend void cruceBlx(Especimen &padreA, Especimen &padreB,float alpha = 0.5);
-    friend void mutar(Especimen &individuo,int probabilidad = 0.1);
+    friend void mutar(Especimen &individuo,double probabilidad = 0.1);
 private:
     std::vector<Transistor>* transistors_;
     std::vector<Restriction>* restrictions_;
